refactor (genderize)
rewrite all promise chains to async-await; add promise-based readline submodule; use URL object with params for endpoint building; divide modules.js to separate modules; change exit command; style: remove unnecessary semicolons
