+++ miamoremustachio/genderize/modules/constants.js
@@ -1,0 +1,5 @@
+const EXIT_COMMAND = 'q';
+
+module.exports = {
+    EXIT_COMMAND,
+};
